requirements.txt
pypdf
google-generativeai
fastapi
gunicorn
uvicorn
jinja2
sentry-sdk[fastapi]
python-docx
reportlab